EVL-145: repaint only the results region, not the whole page
The page component subscribed to the `list` and `lookup` slices, so every
response re-rendered the title, the count and the entire filter bar along with
the rows. v1 never did that — re-frame repaints only the subtree that is
subscribed — and it is what made the screen flash on each request.

Split the two pieces that depend on fetched data into their own subscribers:
`sections/SessionsResults` (loader / empty / error / list / load-more) and
`sections/SessionsCount`. The page itself now subscribes to nothing but the
stable `fetchList` action, so it re-renders only when the URL params change,
which is exactly when the filter bar genuinely needs to update.

SessionsCount also keeps the previous numbers on screen while a refetch is in
flight, so the header does not collapse and reflow the page under the user.
